- added typecasts

+++ dlls/CBotCS.cpp
@@ -190,7 +190,7 @@ void CBotCS :: Init(void){
 	f_RWKnife = gpGlobals->time + 120.0;	// just some time
 	f_blinded_time=0;
 	f_gotohostage = 0;
-	f_Bored = gpGlobals->time + RANDOM_LONG(-TIMETOBEBORED*2/3,TIMETOBEBORED/2);
+	f_Bored = gpGlobals->time + RANDOM_LONG(long(-TIMETOBEBORED*2/3),long(TIMETOBEBORED/2));
 	//f_Hide = 0;
 	//	pHidefrom = 0;
 	//	f_GotoHidingPlace = 0;
@@ -1816,8 +1816,8 @@ void CBotCS :: Think(void){
 	{
 		HandleMenu();
 			
-		g_engfuncs.pfnRunPlayerMove( pEdict, pEdict->v.v_angle, 0.0,
-			0, 0, pEdict->v.button, 0, g_msecval);
+		g_engfuncs.pfnRunPlayerMove( pEdict, pEdict->v.v_angle, 0,
+			0, 0, pEdict->v.button, 0, byte(g_msecval));
 		
 		
 		FixIdealYaw();				// this fixes a bug, which does onl occurr in de_dust, cs_gamma_assault etc.
@@ -1848,7 +1848,7 @@ void CBotCS :: Think(void){
 		lButton = IN_ATTACK;*/
 		
 		g_engfuncs.pfnRunPlayerMove( pEdict, pEdict->v.v_angle, f_move_speed,
-			0, 0, lButton, 0, g_msecval);
+			0, 0, lButton, 0, byte(g_msecval));
 		
 		//cout << "deadthinkend" << endl;
 		return;
@@ -2430,7 +2430,7 @@ void CBotCS :: Think(void){
 	pEdict->v.button = lButton;			// copy the new pressed buttons to the right location
 	
 	g_engfuncs.pfnRunPlayerMove( pEdict, pEdict->v.v_angle, f_move_speed,
-		f_strafe, 0, pEdict->v.button, 0, g_msecval);
+		f_strafe, 0, pEdict->v.button, 0, byte(g_msecval));
 	
 	//BotFixIdealYaw(pEdict);
